Upgrade repo ruby version to 2.2.2
**What does this do?**
It upgrades the ruby version used by the chef-repo project to 2.2.2.

**Why is this needed?**
I'm seting up a new IC server, which is running 2.2.2. So we can finally upgrade
the chef-repo to 2.2.2 as well. It is needed because its a good practice to be
on latest stable ruby.

**Related commits/issues?**
We have been on ruby 2.2.0 in version 2.3.0, but this was downgraded in 2.4.0
(https://github.com/intercity/chef-repo/pull/174)

Signed-off-by: Jeroen van Baarsen <jeroenvanbaarsen@gmail.com>